fmt, clippy

+++ src/tui/ui/pages/modbus_panel/input/navigation.rs
@@ -82,8 +82,7 @@ pub fn handle_navigation_input(key: KeyEvent, bus: &Bus) -> Result<()> {
                             if let Ok(port_guard) = port_entry.read() {
                                 let types::port::PortConfig::Modbus { mode: _, stations } =
                                     &port_guard.config;
-                                let all_items: Vec<_> =
-                                    stations.iter().collect();
+                                let all_items: Vec<_> = stations.iter().collect();
                                 if let Some(item) = all_items.get(slave_index) {
                                     return Ok(item.register_length as usize);
                                 }
@@ -147,8 +146,7 @@ pub fn handle_navigation_input(key: KeyEvent, bus: &Bus) -> Result<()> {
                             if let Ok(port_guard) = port_entry.read() {
                                 let types::port::PortConfig::Modbus { mode: _, stations } =
                                     &port_guard.config;
-                                let all_items: Vec<_> =
-                                    stations.iter().collect();
+                                let all_items: Vec<_> = stations.iter().collect();
                                 if let Some(item) = all_items.get(slave_index) {
                                     return Ok((
                                         item.register_length as usize,
@@ -242,8 +240,7 @@ pub fn handle_navigation_input(key: KeyEvent, bus: &Bus) -> Result<()> {
                             if let Ok(port_guard) = port_entry.read() {
                                 let types::port::PortConfig::Modbus { mode: _, stations } =
                                     &port_guard.config;
-                                let all_items: Vec<_> =
-                                    stations.iter().collect();
+                                let all_items: Vec<_> = stations.iter().collect();
                                 if let Some(item) = all_items.get(slave_index) {
                                     let has_next = slave_index + 1 < all_items.len();
                                     return Ok((
@@ -257,7 +254,7 @@ pub fn handle_navigation_input(key: KeyEvent, bus: &Bus) -> Result<()> {
                         Ok((0, 0, false))
                     })?;
 
-                    let max_row = (max_register as usize).div_ceil(8);
+                    let max_row = max_register.div_ceil(8);
                     let target_index = (row + 1) * 8 + col;
 
                     if target_index < max_register {
